+++ src/ipl3checksum/__main__.py
@@ -17,6 +17,7 @@ def ipl3checksumMain():
 
     subparsers = parser.add_subparsers(description="action", help="The CLI utility to run", required=True)
 
+    ipl3checksum.frontends.check.addSubparser(subparsers)
     ipl3checksum.frontends.detect_cic.addSubparser(subparsers)
     ipl3checksum.frontends.sum.addSubparser(subparsers)